refactor: Enhance cross-compilation support and platform-specific configurations in build scripts

+++ scripts/17-libbluray.sh
@@ -2,20 +2,40 @@
 
 # libgpg-error
 cd /build/libgpg-error
+if [[ ${TARGET_OS} == "darwin" ]]; then
+    if [[ ${ARCH} == "arm64" ]]; then
+        cp src/syscfg/lock-obj-pub.${ARCH%64}-apple-darwin.h src/syscfg/lock-obj-pub.darwin24.1.h
+    else
+        cp src/syscfg/lock-obj-pub.${ARCH}-apple-darwin.h src/syscfg/lock-obj-pub.${CROSS_PREFIX%-}.h
+    fi
+fi
+
 ./autogen.sh --prefix=${PREFIX} --enable-static --disable-shared --with-pic --disable-doc \
     --host=${CROSS_PREFIX%-}
 ./configure --prefix=${PREFIX} --enable-static --disable-shared --with-pic --disable-doc \
-    --host=${CROSS_PREFIX%-}
+    --host=${CROSS_PREFIX%-} | tee /ffmpeg_build.log
+if [ ${PIPESTATUS[0]} -ne 0 ]; then
+    echo "libgpg-error configure failed" >>/ffmpeg_build.log
+    exit 1
+fi
 make -j$(nproc) && make install
 echo "Libs.private: -lstdc++" >>${PREFIX}/lib/pkgconfig/libgpg-error.pc
 rm -rf /build/libgpg-error
 
 # libgcrypt
+EXTRA_FLAGS=""
 cd /build/libgcrypt
-./autogen.sh --prefix=${PREFIX} --enable-static --disable-shared --with-pic --disable-doc \
+if [[ ${TARGET_OS} == "darwin" ]]; then
+    EXTRA_FLAGS="--disable-asm --disable-test"
+fi
+./autogen.sh --prefix=${PREFIX} --enable-static --disable-shared --with-pic --disable-doc ${EXTRA_FLAGS} \
     --host=${CROSS_PREFIX%-}
-./configure --prefix=${PREFIX} --enable-static --disable-shared --with-pic --disable-doc \
-    --host=${CROSS_PREFIX%-}
+./configure --prefix=${PREFIX} --enable-static --disable-shared --with-pic --disable-doc ${EXTRA_FLAGS} \
+    --host=${CROSS_PREFIX%-} | tee /ffmpeg_build.log
+if [ ${PIPESTATUS[0]} -ne 0 ]; then
+    echo "libgcrypt configure failed" >>/ffmpeg_build.log
+    exit 1
+fi
 make -j$(nproc) && make install
 echo "Libs.private: -lstdc++" >>${PREFIX}/lib/pkgconfig/libgcrypt.pc
 
@@ -33,7 +53,11 @@ cd /build/libbdplus
 ./bootstrap --prefix=${PREFIX} --libdir=${PREFIX}/lib --enable-static --disable-shared --with-pic --disable-doc \
     --host=${CROSS_PREFIX%-}
 ./configure --prefix=${PREFIX} --libdir=${PREFIX}/lib --enable-static --disable-shared --with-pic --disable-doc \
-    --host=${CROSS_PREFIX%-}
+    --host=${CROSS_PREFIX%-} | tee /ffmpeg_build.log
+if [ ${PIPESTATUS[0]} -ne 0 ]; then
+    echo "libbdplus configure failed" >>/ffmpeg_build.log
+    exit 1
+fi
 make -j$(nproc) && make install
 echo "Libs.private: -lstdc++" >>${PREFIX}/lib/pkgconfig/libbdplus.pc
 rm -rf /build/libbdplus
@@ -43,13 +67,17 @@ cd /build/libaacs
 ./bootstrap --prefix=${PREFIX} --libdir=${PREFIX}/lib --enable-static --disable-shared --with-pic --disable-doc \
     --host=${CROSS_PREFIX%-}
 ./configure --prefix=${PREFIX} --libdir=${PREFIX}/lib --enable-static --disable-shared --with-pic --disable-doc \
-    --host=${CROSS_PREFIX%-}
+    --host=${CROSS_PREFIX%-} | tee /ffmpeg_build.log
+if [ ${PIPESTATUS[0]} -ne 0 ]; then
+    echo "libaacs configure failed" >>/ffmpeg_build.log
+    exit 1
+fi
 make -j$(nproc) && make install
 echo "Libs.private: -lstdc++" >>${PREFIX}/lib/pkgconfig/libaacs.pc
 rm -rf /build/libaacs
 
 # libbluray
-export EXTRA_LIBS="-L${PREFIX}/lib -laacs -lbdplus"
+EXTRA_LIBS="-L${PREFIX}/lib -laacs -lbdplus"
 cd /build/libbluray
 sed -i 's/dec_init/libbluray_dec_init/g' src/libbluray/disc/dec.c
 sed -i 's/dec_init/libbluray_dec_init/g' src/libbluray/disc/dec.h
@@ -63,7 +91,6 @@ cd /build/libbluray
     --disable-doxygen-doc --disable-doxygen-dot --disable-doxygen-html --disable-doxygen-ps --disable-doxygen-pdf --disable-examples --disable-bdjava-jar \
     --host=${CROSS_PREFIX%-} \
     LIBS="-laacs -lbdplus" | tee /ffmpeg_build.log
-
 if [ ${PIPESTATUS[0]} -ne 0 ]; then
     exit 1
 fi
@@ -71,7 +98,7 @@ fi
 make -j$(nproc) && make install
 echo "Libs.private: -laacs -lbdplus -lstdc++" >>${PREFIX}/lib/pkgconfig/libbluray.pc
 rm -rf /build/libbluray
-export EXTRA_LIBS=""
+EXTRA_LIBS=""
 
 add_enable "--enable-libbluray"
